pythonProject/requirements.txt
# -*- coding: utf-8 -*-
pyyaml
matplotlib
numpy
